build: tokenomics and initial contract parameters initialization script, refactoring folders, hiding unused files
