+++ addresses.ts
@@ -3,9 +3,9 @@
 export const FUJI_DAI_DEFO_WHALE_ADDRESS = "0x83c4c6F0Ac462481786b8325f1e291c8f456CB00";
 
 export const FUJI_DAI_ADDRESS = "0x85a2ff500E0eD9fA93719071EA46A86198181581";
-export const FUJI_DEFO_ADDRESS = ""; //not deployed yet
+export const FUJI_DEFO_ADDRESS = null; //not deployed yet
 export const FUJI_JOE_ROUTER_ADDRESS = "0xd7f655E3376cE2D7A2b08fF01Eb3B1023191A901";
 
 export const MAINNET_DAI_ADDRESS = "0xd586e7f844cea2f87f50152665bcbc2c279d8d70";
-export const MAINNET_DEFO_ADDRESS = ""; //not deployed yet
+export const MAINNET_DEFO_ADDRESS = null; //not deployed yet
 export const MAINNET_JOE_ROUTER_ADDRESS = "0x60aE616a2155Ee3d9A68541Ba4544862310933d4";
