+++ constants/index.ts
@@ -1,2 +1,4 @@
 export const SECONDS_IN_A_DAY = 86400;
 export const ONE_YEAR_IN_SEC = SECONDS_IN_A_DAY * 365;
+
+export * from "./config";
