+++ deploy/999_post_checks.ts
@@ -1,15 +1,12 @@
 import { DeployFunction } from "hardhat-deploy/types";
 import { HardhatRuntimeEnvironment } from "hardhat/types";
 
-import { deploySuccess } from "../utils/helpers";
+import { deployInfo } from "../utils/helpers";
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
-  const { deployments, getNamedAccounts } = hre;
-  const { read } = deployments;
-  const { deployer } = await getNamedAccounts();
-
-  const defoToken = await deployments.get("DEFOToken");
-  deploySuccess(`DEFOToken verified at ${defoToken.address} : ${await read("DEFOToken", "balanceOf", deployer)}`);
+  const { deployments } = hre;
+  await deployments.get("DEFOToken");
+  deployInfo("Verified");
 };
 
 export default func;
