+++ hardhat.config.ts
@@ -64,6 +64,10 @@ const config: HardhatUserConfig = {
       },
     ],
   },
+  typechain: {
+    outDir: "types",
+    target: "ethers-v5",
+  },
 };
 
 runtimeTasks;
