+++ unused/Stablisier/helper/addresses.js
@@ -1,0 +1,5 @@
+exports.routerAddress = "0x8Ba0297709D77CdFBdD35574aD60d5D4F5A92257";
+exports.factoryAddress = "0x3141110EDbf0c16c338F1e9A38570D0abD368064";
+exports.lpAddress = "0x73F96647d4552f52AE07880cC61e5BEc2D14346B";
+exports.defoAddress = "0xE99cf700591F3605ad4494E166D2f447F0F50a71";
+exports.mDaiAddress = "0xEEaE62ce626c6A27b431A7f9670B7F209ab148F7"
