+++ unused/scripts/libraries/diamond.js
