+++ unused/test/node-diamond-test.js
