+++ unused/test/redeem-test.js
